Make InfoBanner2 the only InfoBanner
