+++ common/views/components/DefaultPageLayout/DefaultPageLayout.js
@@ -4,7 +4,7 @@ import Head from 'next/head';
 import NastyJs from '../Header/NastyJs';
 import Header from '../Header/Header';
 import NewsletterPromo from '../NewsletterPromo/NewsletterPromo';
-import InfoBanner from '../InfoBanner/InfoBanner2';
+import InfoBanner from '../InfoBanner/InfoBanner';
 import {striptags} from '../../../utils/striptags';
 import {formatDate} from '../../../utils/format-date';
 import {museumLd, objToJsonLd} from '../../../utils/json-ld';
