--- common/views/components/InfoBanner/InfoBanner.config.js
@@ -1,4 +1,0 @@
-export const name = 'Info Banner';
-export const context = {
-  text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec volutpat dictum nisi vel malesuada. Aenean consectetur pretium sodales. In convallis nibh enim, at euismod justo bibendum id.'
-};
